+++ R/Create.actor.mastodon.R
@@ -5,6 +5,8 @@
 #'
 #' @param datasource Collected social media data with \code{"datasource"} and \code{"mastodon"} class names.
 #' @param type Character string. Type of network to be created, set to \code{"actor"}.
+#' @param subtype Character string. Subtype of actor network to be created. Can be set to \code{"server"}. Default is
+#'   \code{NULL}.
 #' @param inclMentions Logical. Create edges for users mentioned or tagged in posts. Default is \code{TRUE}.
 #' @param verbose Logical. Output additional information about the network creation. Default is \code{FALSE}.
 #' @param ... Additional parameters passed to function. Not used in this method.
@@ -15,12 +17,16 @@
 #' \dontrun{
 #' # create a mastodon actor network
 #' actor_net <- mastodon_data |> Create("actor")
+#'
+#' # create a mastodon server relations network
+#' actor_net <- mastodon_data |> Create("actor", "server")
 #' }
 #'
 #' @export
 Create.actor.mastodon <-
   function(datasource,
            type,
+           subtype = NULL,
            inclMentions = TRUE,
            verbose = FALSE,
            ...) {
@@ -31,59 +37,109 @@ Create.actor.mastodon <-
       stop("Datasource invalid or empty.", call. = FALSE)
     }
     
+    check_chr(subtype, param = "subtype", accept = "server", null.ok = TRUE)
+    
     relations <- datasource$posts |>
-      dplyr::select(post.id = .data$id, .data$account, .data$in_reply_to_account_id, .data$mentions,
-                    .data$created_at) |>
-      tidyr::hoist(.data$account,
-                   account.id = "id",
-                   account.acct = "acct",
-                   account.username = "username",
-                   account.displayname = "display_name") |>
-      dplyr::select(-.data$account) |>
+      dplyr::select(
+        post.id = "id",
+        post.created_at = "created_at",
+        "account",
+        "in_reply_to_account_id",
+        "mentions"
+      ) |>
+      tidyr::hoist(
+        "account",
+        user.id = "id",
+        user.acct = "acct",
+        user.username = "username",
+        user.displayname = "display_name",
+        user.url = "url",
+        user.avatar = "avatar"
+      ) |>
+      dplyr::select(-"account") |>
       dplyr::mutate(edge_type = "reply")
     
     edges <- relations |>
-      dplyr::select(from = .data$account.id, to = .data$in_reply_to_account_id, .data$post.id, .data$created_at,
-                    .data$edge_type)
+      dplyr::select(
+        from = "user.id",
+        to = "in_reply_to_account_id",
+        "post.id",
+        "post.created_at",
+        "edge_type")
     
+    nodes <- relations |>
+      dplyr::select(dplyr::starts_with("user.")) |>
+      dplyr::distinct(.data$user.id, .keep_all = TRUE) |>
+      dplyr::mutate(type = "author")
+
     if (inclMentions) {
       mentions <- relations |>
-        dplyr::select(.data$post.id, .data$account.id, .data$in_reply_to_account_id, .data$mentions,
-                      .data$created_at) |>
-        tidyr::hoist(.data$mentions,
-                     mention.account.id = list("id"),
-                     mention.account.acct = list("acct"),
-                     mention.account.username = list("username"),
-                     mention.account.displayname = list("display_name")) |>
-        tidyr::unnest_longer(
-          c(.data$mention.account.id,
-            .data$mention.account.acct,
-            .data$mention.account.username,
-            .data$mention.account.displayname)
+        tidyr::hoist(
+          "mentions",
+          mention.user.id = list("id"),
+          mention.user.acct = list("acct"),
+          mention.user.username = list("username"),
+          mention.user.url = list("url")
         ) |>
-        dplyr::filter(.data$mention.account.id != .data$in_reply_to_account_id) |> # mention is already a direct reply
+        tidyr::unnest_longer(
+          c(
+            "mention.user.id",
+            "mention.user.acct",
+            "mention.user.username",
+            "mention.user.url"
+          )
+        ) |>
+        dplyr::filter(.data$mention.user.id != .data$in_reply_to_account_id) |> # mention is already a direct reply
         dplyr::mutate(edge_type = "mention")
       
       edges <- edges |>
         dplyr::bind_rows(
           mentions |>
-            dplyr::select(from = .data$account.id,
-                          to = .data$mention.account.id,
-                          .data$post.id,
-                          .data$created_at,
-                          .data$edge_type)
+            dplyr::select(
+              from = "user.id",
+              to = "mention.user.id",
+              "post.id",
+              "post.created_at",
+              "edge_type"
+            )
         )
+      
+      nodes <- nodes |>
+        dplyr::bind_rows(
+          mentions |>
+            dplyr::select(dplyr::starts_with("mention.user.")) |>
+            dplyr::rename_with(~ gsub("mention.", "", .x, fixed = TRUE)) |>
+            dplyr::mutate(type = "mention")
+        ) |>
+        dplyr::arrange("user.id", "type") |>
+        dplyr::distinct(.data$user.id, .keep_all = TRUE)
+    }
+
+    # add referenced actors absent from nodes
+    nodes <- add_absent_nodes(nodes, edges, id = "user.id")
+    
+    if (!is.null(subtype)) {
+      servers <- nodes |>
+        dplyr::select("user.id", "user.url") |>
+        dplyr::rowwise() |>
+        dplyr::mutate(user.server = list(httr::parse_url(.data$user.url)$hostname)) |>
+        dplyr::ungroup()
+      
+      edges <- edges |>
+        dplyr::left_join(servers |> dplyr::select("user.id", from.server = "user.server"), by = c("from" = "user.id")) |>
+        dplyr::left_join(servers |> dplyr::select("user.id", to.server = "user.server"), by = c("to" = "user.id")) |>
+        dplyr::select(from = "from.server", to = "to.server", "edge_type") |>
+        dplyr::filter(!is.null(.data$to) & .data$to != "NULL")
+      
+      nodes <- servers |>
+        dplyr::group_by(.data$user.server) |>
+        dplyr::summarise(n = dplyr::n())
     }
     
     edges <- edges |> dplyr::filter(!is.na(.data$to))
     
-    nodes <- datasource$users |> dplyr::mutate(node_type = "user")
-    
-    # add referenced actors absent from nodes
-    nodes <- add_absent_nodes(nodes, edges)
-    
     net <- list("edges" = edges, "nodes" = nodes)
-    class(net) <- append(c("network", "actor", "mastodon"), class(net))
+    class(net) <- append(c("network", "actor", switch(!is.null(subtype), "server", NULL), "mastodon"), class(net))
     msg("Done.\n")
     
     net