+++ R/vosonSML-package.R
@@ -3,9 +3,9 @@
 #' @docType package
 #' @keywords internal
 "_PACKAGE"
-#' @importFrom data.table data.table setkey ":="
+#' @importFrom data.table data.table setkey :=
 #' @importFrom methods new
-#' @importFrom rlang ".data" check_installed
+#' @importFrom rlang .data check_installed
 #' @importFrom stats na.omit runif setNames
 #' @importFrom utils capture.output flush.console read.table tail
 utils::globalVariables(c("collect_log", "merge_log"))